+++ src/main/java/io/zerogone/blog/repository/BlogDao.java
@@ -1,7 +1,10 @@
 package io.zerogone.blog.repository;
 
 import io.zerogone.blog.model.Blog;
+import io.zerogone.blogmember.model.BlogMember;
+import io.zerogone.blogmember.model.MemberRole;
 import io.zerogone.exception.UniquePropertyException;
+import io.zerogone.user.model.User;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.springframework.stereotype.Repository;
@@ -9,6 +12,12 @@ import org.springframework.stereotype.Repository;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import javax.persistence.PersistenceException;
+import javax.persistence.TypedQuery;
+import javax.persistence.criteria.CriteriaBuilder;
+import javax.persistence.criteria.CriteriaQuery;
+import javax.persistence.criteria.Join;
+import javax.persistence.criteria.Root;
+import java.util.List;
 
 @Repository
 public class BlogDao {
@@ -29,5 +38,23 @@ public class BlogDao {
 
         logger.debug("blog id: " + blog.getId());
         logger.debug("-----save blog end-----");
+    }
+
+    public List<Blog> findAllByUser(User user) {
+        CriteriaBuilder criteriaBuilder = entityManager.getCriteriaBuilder();
+        CriteriaQuery<Blog> criteriaQuery = criteriaBuilder.createQuery(Blog.class);
+
+        Root<Blog> root = criteriaQuery.from(Blog.class);
+        Join<Blog, BlogMember> join = root.join("members");
+
+        criteriaQuery.select(root);
+        criteriaQuery.where(criteriaBuilder.and(
+                criteriaBuilder.equal(join.get("userId"), user.getId()),
+                criteriaBuilder.or(
+                        criteriaBuilder.equal(join.get("role"), MemberRole.MEMBER),
+                        criteriaBuilder.equal(join.get("role"), MemberRole.ADMIN))));
+
+        TypedQuery<Blog> typedQuery = entityManager.createQuery(criteriaQuery);
+        return typedQuery.getResultList();
     }
 }
