+++ src/main/java/io/zerogone/blog/service/BlogSearchService.java
@@ -1,21 +1,22 @@
 package io.zerogone.blog.service;
 
 import io.zerogone.blog.model.Blog;
-import io.zerogone.blog.repository.BlogSearchDao;
-import io.zerogone.model.User;
+import io.zerogone.blog.repository.BlogDao;
+import io.zerogone.user.model.CurrentUserInfo;
+import io.zerogone.user.model.User;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 @Service
 public class BlogSearchService {
-    private final BlogSearchDao blogSearchDao;
+    private final BlogDao blogDao;
 
-    public BlogSearchService(BlogSearchDao blogSearchDao) {
-        this.blogSearchDao = blogSearchDao;
+    public BlogSearchService(BlogDao blogDao) {
+        this.blogDao = blogDao;
     }
 
-    public List<Blog> getBlogsThatUserBelongTo(User user) {
-        return blogSearchDao.findAllByUser(user);
+    public List<Blog> getBlogsThatUserBelongTo(CurrentUserInfo userInfo) {
+        return blogDao.findAllByUser(new User(userInfo));
     }
 }
